refactor(app): split workspace/session sidebar into component
Move the sidebar workspace/session list out of session.tsx so session view logic stays focused and future long-session iterations can evolve independently without reworking a 4k-line page.
